pyproject.toml
[project]
name = "whatsthatfish"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "albumentations>=1.4.0",
    "boto3>=1.42.78",
    "ensemble-boxes>=1.0.8",
    "gcsfs>=2024.2.0",
    "google-cloud-storage>=3.10.1",
    "numpy>=1.24.0",
    "opencv-python-headless>=4.9.0",
    "pillow>=10.0.0",
    "pycocotools>=2.0.7",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0",
    "scikit-image>=0.22.0",
    "supervision>=0.19.0",
    "tenacity>=9.1.4",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "tqdm>=4.66.0",
    "ultralytics>=8.3.0",
]
